Implemented alarm panel and high pressure alarm

+++ ApolloBVM/src/nscdrrn001pdunv.h
@@ -13,8 +13,8 @@ class Nscdrrn001pdunv {
   public:
     Nscdrrn001pdunv(uint8_t pin, float pressure_conversion_multiplier, float pressure_conversion_constant );
 
-    float read();
-    float convertToPressure(int input);
+    int read();
+    int convertToPressure(int input);
 
 
   private:
